client/cli: Organize files under a 'cli' directory
- Also includes `clang-format` changes
