+++ src/client/cli/CMakeLists.txt
@@ -12,6 +12,37 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-add_subdirectory(cli)
-add_subdirectory(common)
-add_subdirectory(gui)
+add_library(argparser STATIC
+  argparser.cpp)
+
+target_link_libraries(argparser
+  fmt
+  rpc
+  Qt5::Core
+  Qt5::Network)
+
+add_library(client STATIC
+  client.cpp)
+
+target_link_libraries(client
+  argparser
+  cert
+  commands
+  fmt
+  formatter
+  platform
+  rpc
+  Qt5::Core)
+
+add_executable(multipass
+  main.cpp)
+
+target_link_libraries(multipass
+  client)
+
+install(TARGETS multipass
+  DESTINATION bin
+  COMPONENT multipass)
+
+add_subdirectory(cmd)
+add_subdirectory(formatter)
