+++ src/client/cli/cmd/common_cli.h
