+++ src/client/cli/cmd/delete.cpp
