+++ src/client/cli/cmd/delete.h
