+++ src/client/cli/cmd/find.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017-2018 Canonical, Ltd.
+ * Copyright (C) 2017 Canonical, Ltd.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -13,37 +13,32 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
- * Authored by: Alberto Aguirre <alberto.aguirre@canonical.com>
- *
  */
 
-#ifndef MULTIPASS_LIST_H
-#define MULTIPASS_LIST_H
+#ifndef MULTIPASS_FIND_H
+#define MULTIPASS_FIND_H
 
 #include <multipass/cli/command.h>
 
 namespace multipass
 {
-class Formatter;
-
 namespace cmd
 {
-class List final : public Command
+class Find final : public Command
 {
 public:
     using Command::Command;
-    ReturnCode run(ArgParser *parser) override;
+    ReturnCode run(ArgParser* parser) override;
 
     std::string name() const override;
-    std::vector<std::string> aliases() const override;
     QString short_help() const override;
     QString description() const override;
 
 private:
-    ParseCode parse_args(ArgParser *parser) override;
+    FindRequest request;
 
-    Formatter* chosen_formatter;
+    ParseCode parse_args(ArgParser* parser) override;
 };
-}
-}
-#endif // MULTIPASS_LIST_H
+} // namespace cmd
+} // namespace multipass
+#endif // MULTIPASS_FIND_H
