+++ src/client/cli/cmd/info.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017 Canonical, Ltd.
+ * Copyright (C) 2017-2018 Canonical, Ltd.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -13,33 +13,37 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
+ * Authored by: Chris Townsend <christopher.townsend@canonical.com>
+ *
  */
 
-#ifndef MULTIPASS_UMOUNT_H
-#define MULTIPASS_UMOUNT_H
+#ifndef MULTIPASS_INFO_H
+#define MULTIPASS_INFO_H
 
 #include <multipass/cli/command.h>
 
 namespace multipass
 {
+class Formatter;
+
 namespace cmd
 {
-class Umount final : public Command
+class Info final : public Command
 {
 public:
     using Command::Command;
-    ReturnCode run(ArgParser *parser) override;
+    ReturnCode run(ArgParser* parser) override;
 
     std::string name() const override;
-    std::vector<std::string> aliases() const override;
     QString short_help() const override;
     QString description() const override;
 
 private:
-    UmountRequest request;
+    InfoRequest request;
+    Formatter* chosen_formatter;
 
-    ParseCode parse_args(ArgParser *parser) override;
+    ParseCode parse_args(ArgParser* parser) override;
 };
-}
-}
-#endif // MULTIPASS_UMOUNT_H
+} // namespace cmd
+} // namespace multipass
+#endif // MULTIPASS_INFO_H
