+++ src/client/cli/cmd/list.cpp
