+++ src/client/cli/cmd/purge.cpp
