+++ src/client/cli/cmd/recover.cpp
@@ -15,17 +15,16 @@
  *
  */
 
-#include "info.h"
+#include "recover.h"
 #include "common_cli.h"
 
 #include <multipass/cli/argparser.h>
-#include <multipass/cli/formatter.h>
 
 namespace mp = multipass;
 namespace cmd = multipass::cmd;
 using RpcMethod = mp::Rpc::Stub;
 
-mp::ReturnCode cmd::Info::run(mp::ArgParser* parser)
+mp::ReturnCode cmd::Recover::run(mp::ArgParser* parser)
 {
     auto ret = parse_args(parser);
     if (ret != ParseCode::Ok)
@@ -33,56 +32,45 @@ mp::ReturnCode cmd::Info::run(mp::ArgParser* parser)
         return parser->returnCodeFrom(ret);
     }
 
-    auto on_success = [this](mp::InfoReply& reply) {
-        cout << chosen_formatter->format(reply);
-
-        return ReturnCode::Ok;
-    };
+    auto on_success = [](mp::RecoverReply& reply) { return mp::ReturnCode::Ok; };
 
     auto on_failure = [this](grpc::Status& status) { return standard_failure_handler_for(name(), cerr, status); };
 
     request.set_verbosity_level(parser->verbosityLevel());
-    return dispatch(&RpcMethod::info, request, on_success, on_failure);
+    return dispatch(&RpcMethod::recover, request, on_success, on_failure);
 }
 
-std::string cmd::Info::name() const { return "info"; }
-
-QString cmd::Info::short_help() const
+std::string cmd::Recover::name() const
 {
-    return QStringLiteral("Display information about instances");
+    return "recover";
 }
 
-QString cmd::Info::description() const
+QString cmd::Recover::short_help() const
 {
-    return QStringLiteral("Display information about instances");
+    return QStringLiteral("Recover deleted instances");
 }
 
-mp::ParseCode cmd::Info::parse_args(mp::ArgParser* parser)
+QString cmd::Recover::description() const
 {
-    parser->addPositionalArgument("name", "Names of instances to display information about", "<name> [<name> ...]");
+    return QStringLiteral("Recover deleted instances so they can be used again.");
+}
 
-    QCommandLineOption all_option(all_option_name, "Display info for all instances");
+mp::ParseCode cmd::Recover::parse_args(mp::ArgParser* parser)
+{
+    parser->addPositionalArgument("name", "Names of instances to recover", "<name> [<name> ...]");
+
+    QCommandLineOption all_option(all_option_name, "Recover all deleted instances");
     parser->addOption(all_option);
 
-    QCommandLineOption formatOption(
-        "format", "Output info in the requested format.\nValid formats are: table (default), json, csv and yaml",
-        "format", "table");
-    parser->addOption(formatOption);
-
     auto status = parser->commandParse(this);
-
     if (status != ParseCode::Ok)
-    {
         return status;
-    }
 
     auto parse_code = check_for_name_and_all_option_conflict(parser, cerr);
     if (parse_code != ParseCode::Ok)
         return parse_code;
 
     request.mutable_instance_names()->CopyFrom(add_instance_names(parser));
-
-    status = handle_format_option(parser, &chosen_formatter, cerr);
 
     return status;
 }
