+++ src/client/cli/cmd/restart.cpp
