+++ src/client/cli/cmd/umount.cpp
@@ -32,9 +32,7 @@ mp::ReturnCode cmd::Umount::run(mp::ArgParser* parser)
         return parser->returnCodeFrom(ret);
     }
 
-    auto on_success = [](mp::UmountReply& reply) {
-        return ReturnCode::Ok;
-    };
+    auto on_success = [](mp::UmountReply& reply) { return ReturnCode::Ok; };
 
     auto on_failure = [this](grpc::Status& status) { return standard_failure_handler_for(name(), cerr, status); };
 
@@ -42,7 +40,10 @@ mp::ReturnCode cmd::Umount::run(mp::ArgParser* parser)
     return dispatch(&RpcMethod::umount, request, on_success, on_failure);
 }
 
-std::string cmd::Umount::name() const { return "umount"; }
+std::string cmd::Umount::name() const
+{
+    return "umount";
+}
 
 std::vector<std::string> cmd::Umount::aliases() const
 {
@@ -61,9 +62,10 @@ QString cmd::Umount::description() const
 
 mp::ParseCode cmd::Umount::parse_args(mp::ArgParser* parser)
 {
-    parser->addPositionalArgument("mount", "Mount points, in <name>[:<path>] format, where <name> "
-                                           "are instance names, and optional <path> are mount points. "
-                                           "If omitted, all mounts will be removed from the named instances.",
+    parser->addPositionalArgument("mount",
+                                  "Mount points, in <name>[:<path>] format, where <name> "
+                                  "are instance names, and optional <path> are mount points. "
+                                  "If omitted, all mounts will be removed from the named instances.",
                                   "<mount> [<mount> ...]");
 
     auto status = parser->commandParse(this);
