+++ src/client/cli/cmd/version.cpp
@@ -17,6 +17,7 @@
 
 #include "version.h"
 #include "common_cli.h"
+
 #include <multipass/cli/argparser.h>
 #include <multipass/version.h>
