+++ src/client/cli/cmd/version.h
@@ -17,8 +17,8 @@
  *
  */
 
-#ifndef MULTIPASS_CONNECT_H
-#define MULTIPASS_CONNECT_H
+#ifndef MULTIPASS_VERSION_H
+#define MULTIPASS_VERSION_H
 
 #include <multipass/cli/command.h>
 
@@ -26,22 +26,19 @@ namespace multipass
 {
 namespace cmd
 {
-class Shell final : public Command
+class Version final : public Command
 {
 public:
     using Command::Command;
-    ReturnCode run(ArgParser *parser) override;
+    ReturnCode run(ArgParser* parser) override;
 
     std::string name() const override;
-    std::vector<std::string> aliases() const override;
     QString short_help() const override;
     QString description() const override;
 
 private:
-    SSHInfoRequest request;
-
-    ParseCode parse_args(ArgParser *parser) override;
+    ParseCode parse_args(ArgParser* parser) override;
 };
-}
-}
-#endif // MULTIPASS_CONNECT_H
+} // namespace cmd
+} // namespace multipass
+#endif // MULTIPASS_VERSION_H
