+++ src/client/cli/formatter/format_utils.cpp
