+++ src/client/cli/formatter/json_formatter.cpp
