+++ src/client/cli/formatter/table_formatter.cpp
