+++ tests/test_cli_client.cpp
@@ -23,7 +23,7 @@
 
 #include <multipass/constants.h>
 #include <multipass/logging/log.h>
-#include <src/client/client.h>
+#include <src/client/cli/client.h>
 #include <src/daemon/daemon_rpc.h>
 
 #include <QEventLoop>
